Model class moved to factory
